refactor: added ThemeButton component.
 + refactor: rewrite theme button with composition API instead of Options API.
 + refactor: removed scoped attr in style file.
