+++ src/main/webapp/app/common/primary/header/Header.component.ts
@@ -1,11 +1,13 @@
 import { defineComponent } from 'vue';
 import { IconVue } from '@/common/primary/icon';
+import { ThemeButtonVue } from '@/common/primary/theme-button';
 
 export default defineComponent({
   name: 'Header',
 
   components: {
     IconVue,
+    ThemeButtonVue,
   },
 
   setup() {
@@ -14,46 +16,5 @@ export default defineComponent({
     return {
       selectorPrefix,
     };
-  },
-
-  data() {
-    return {
-      userTheme: 'light-theme',
-    };
-  },
-
-  mounted() {
-    const initUserTheme = (this.getTheme() as 'light-theme' | 'dark-theme') || this.getMediaPreference();
-    this.setTheme(initUserTheme);
-  },
-
-  methods: {
-    toggleTheme() {
-      const activeTheme = localStorage.getItem('user-theme');
-      if (activeTheme === 'light-theme') {
-        this.setTheme('dark-theme');
-      } else {
-        this.setTheme('light-theme');
-      }
-    },
-
-    getTheme() {
-      return localStorage.getItem('user-theme');
-    },
-
-    setTheme(theme: 'light-theme' | 'dark-theme') {
-      localStorage.setItem('user-theme', theme);
-      this.userTheme = theme;
-      document.documentElement.className = theme;
-    },
-
-    getMediaPreference() {
-      const hasDarkPreference: MediaQueryList['matches'] = window.matchMedia('(prefers-color-scheme: dark)').matches;
-      if (hasDarkPreference) {
-        return 'dark-theme';
-      } else {
-        return 'light-theme';
-      }
-    },
   },
 });
