+++ src/main/webapp/app/common/primary/theme-button/index.ts
@@ -1,0 +1,4 @@
+import ThemeButtonComponent from './ThemeButton.component';
+import ThemeButtonVue from './ThemeButton.vue';
+
+export { ThemeButtonComponent, ThemeButtonVue };
